+++ packages/_private/e2e-demo-test/src/PageObjects/ComplexBugs.ts
@@ -1,0 +1,14 @@
+import { PageObject } from './PageObject.js';
+import selectors from '../test.selector';
+
+export class ComplexBugsPage extends PageObject {
+  protected initialPage = PageObject.availablePages.complexBugs;
+
+  async activateItem(locator: keyof typeof selectors) {
+    if (!locator.endsWith('Link')) {
+      throw new Error('Only links can be used for activation!');
+    }
+    const foundLocator = this.getLocator(locator);
+    await foundLocator.click();
+  }
+}